pyproject.toml
[tool.poetry]

name = "airflow_indexima"
version = "2.2.0-b1"
description = "Indexima Airflow integration"


license = "MIT"

authors = ["Jerome Guibert <jguibert@gmail.com>"]

readme = "README.md"

homepage = "https://pypi.org/project/airflow_indexima"
documentation = "https://geronimo-iia.github.io/airflow-indexima/"
repository = "https://github.com/geronimo-iia/airflow-indexima"

keywords = []

classifiers = [
    # update this list to match your application: https://pypi.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed"
]

[tool.poetry.dependencies]

python = "^3.6"
#apache-airflow (>=1.10.3,<2.0.0) requires boto3 (>=1.7.0,<1.8.0).
#boto3 = "1.7.*"
apache-airflow = "1.10.2"
gitpython = "2.1.*"
PyHive="0.6.1"
thrift=">0.10.0"
thrift-sasl="0.3.0"

[tool.poetry.dev-dependencies]

# Formatters
black = "=19.3b0"
isort = "=4.3.4"

# Linters
mypy = "*"
pydocstyle = "*"
flake8 = "*"
flake8-blind-except = "*"
flake8-bugbear = "*"
flake8-builtins = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-docstrings = "*"
flake8-isort = "*"
flake8-black = "*"
flake8-quotes = "*"
flake8-string-format = "*"

# Testing
pytest = "^4.3.1"
pytest-cov = "*"
pytest-describe = "*"
pytest-expecter = "*"
pytest-random = "*"
freezegun = "*"

# Reports
coveragespace = "*"

# Documentation
pydoc-markdown = "2.0.5"
pymdown-extensions = "5.0"
pygments = "*"
mkdocs-material= "*"
fontawesome_markdown = "*"
# needed for pyreverse (https://www.logilab.org/blogentry/6883)
pylint = "*"

# Tooling
pyinstaller = "*"
sniffer = "*"
MacFSEvents = { version = "*", platform = "darwin" }
pync = { version = "*", platform = "darwin" }
autopep8 = "^1.4"


[tool.black]

target-version = ["py37"]
skip-string-normalization = true
line_length = 110

[build-system]

requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.plugins."airflow.plugins"]
"indexima" = "airflow_indexima.indexima:IndeximaAirflowPlugin"
